requirements.txt
# ── API Framework ──────────────────────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# ── Computer Vision ────────────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.1.0
opencv-python-headless>=4.8.0
albumentations>=1.3.1
timm>=0.9.12

# ── LLM / AI ──────────────────────────────────────────────────────────────────
anthropic>=0.34.0

# ── Data Engineering ──────────────────────────────────────────────────────────
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0

# ── Visualization (notebooks / training) ──────────────────────────────────────
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.0
loguru>=0.7.2

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.21.0
